Create common module and move Plugin there
