+++ format/src/main/scala/org/calinburloiu/music/microtonalist/format/JsonSoftChromaticGenusMappingPluginFormat.scala
@@ -16,7 +16,7 @@
 
 package org.calinburloiu.music.microtonalist.format
 
-import org.calinburloiu.music.microtonalist.core.{Plugin, SoftChromaticGenusMapping}
+import org.calinburloiu.music.microtonalist.core.SoftChromaticGenusMapping
 
 object JsonSoftChromaticGenusMappingPluginFormat extends JsonPluginFormat[SoftChromaticGenusMapping] {
